pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.30.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "sift-stack-py>=0.8.3",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "apscheduler>=3.11.0",
    "flask-login>=0.6.3",
]
